Better sizing of preview image, same as for `EncodedImage`
